prov/efa: add efa_fork_support status enum, update fork support logic
Add a new global variable and enum to determine whether fork support is
required. Currently, ON or OFF is valid and UNNEEDED will be added later
when fork support is no longer relevant, those changes are expected to
land in linux kernel version 5.13.

Update the fork handler help text as we support fork again for all
applications, and only install it when fork support is off.

Remove the unsupported fork check as it is no longer needed. Also move
the cache init to its own function while we're here for clarity.

Signed-off-by: Robert Wespetal <wesper@amazon.com>