common/pollfds: Fix possible use after free issue
See #6569

In ofi_pollfds_wait, there are 2 loops that iterate over
the results of poll.  The loops start and stop at a index
value that's maintained between wait calls in order to
ensure some level of fairness in the order that events
are processed.  However, there's a race condition where
an fd can be removed from the pollfds, without that
index being updated.  As a result, the index can be larger
than the current number of fd's maintained by the pollfds.

This makes it possible for pfds->context[i] to reference
freed memory, as pfds->fds[i] was removed by process_work
and is no longer valid.  Simplify the code and use a single
loop that removes the race.

This also fixes an issue where the index is updated in the
first loop, which can result in the second loop iterating
over the same set of fds.  That can report events for the
same fd twice.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>