man/cq: Document completion semantics around HMEM
Greatly expand the semantics around completions and target
side visibility of the data.  No attempt is being made to
define new behavior of the APIs here.  Rather it should be
describing the behavior that applications are already seeing,
but developers may not be fully aware of.

The visibility of transfer at the target is defined, with
additional descriptions for how applications must handle
heterogenous memory.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>