prov/efa: do not use shm provider when application request FI_HMEM
Currently, shm provider is used for communications between buffers
on same host. When the buffer is backed by cuda memory, shm
provider uses the SAR protocol, whose performance is worse than
using device.

This patch implements a short term fix to this issue, which
is to disable shm provider usage when application requested
the FI_HMEM capablity.

It will be reverted when shm provider start to support cuda IPC
communication for cuda buffers on same host.

Signed-off-by: Wei Zhang <wzam@amazon.com>