GitHub Action: Nroff Elves
This GitHub action will fire for every PR.  It will effectively detect
if any man/*.md files have changed, and if so, render new
corresponding nroff man pages in man/man?/*, commit them, and then
push the new commit to that PR.

This is a change compared to the old nroff-auto-generation scheme,
which would see man/*.md changes *after* a PR was merged, and create a
new PR with updated nroff-rendered man pages in man/man?/*, wait for
CI to complete, and then merge that PR.

The rationale for this change is as follows:

1. This GitHub Action is *significantly* simpler than the old scheme.
2. GitHub Actions are both free and easy to maintain.
3. Moving to GitHub Actions for the Nroff Elves removes one more
   service from the flatbed2.openfabric.org server, which moves them
   one step towards eliminating that (expensive) server.

NOTE: The *first* time this GitHub Action runs (which will be *after*
this PR merges to the main libfabric repo), it will cause a
re-generation of *all* the existing man pages.  This is because the
version of Pandoc used to generate the nroff is significantly newer on
the GitHub Action Ubuntu container (compared to the Pandoc version
we're running on flatbed2.openfabric.org).  Hence, the rendered nroff
will change in all files in lots of little, effectively meaningless
ways.  After that, we should see the usual/expected behavior of only
seeing OFIWG Bot commits from the Nroff Elves only when man/*.md files
change.

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>