core: Update dynamic_rbuf op to allow reporting tags
Update ofix_dynamic_rbuf to version 2.  Update the get_rbuf
callback to use an internal cq entry format in place of
fi_cq_data_entry.  The tcp provider will need this to
report tagged messages to rxm, when tcp is carrying the
tag value in its header, and the endpoint context that
a message was received on.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>