Migrate the Nroff Elves to GitHub Actions
This is attempt two: the first attempt didn't work out because GitHub
Actions are really, really not setup to have write permissions on fork
repos creating PRs against the ofiwg/libfabric github repo.

This scheme is schedule-based:

- This GitHub Action will run every 15 minutes (or, more likely given
  our status as a free project at GitHub, at least once an hour)
- It will check out the default branch
- It will convert all the markdown to nroff
- If there are any changes to the generated nroff compared to the
  committed code:
  - Commit the changes
  - Push to a branch on the ofiwg/libfabric repo
  - Make a pull request with the changes
  - Wait for the mandated "DCO" CI to complete
  - Merge the pull request
  - Delete the branch on the ofiwg/libfabric repo

Do all of the above in a relatively short shell script, heavily
utilizing the "hub" command (see https://hub.github.com/) to handle
the majority of interactions with the GitHub API.

The above is essentially the same scheme Libfabric currently uses via
the config/cron-run-all-md2nroff.pl script running via cron on
flatbed2.openfabrics.org.  The GitHub Action version of this scheme,
however, is significantly simpler, mainly because it runs entirely
within the GitHub infrastructure.  The migration to GitHub Actions
means that there is one less reason to need flatbed2.openfabrics.org.

Note that the scheduled invocations of this GitHub Action will *only*
run against the default repository branch (currently "master").  If
this Action needs to be invoked against a release branch, it can be
invoked manually at any time (and it will pick up *all* cumulative
markdown documentation changes on that branch -- there's no need to run
this workflow for each PR that changes the markdown docs):

1. Go to https://github.com/ofiwg/libfabric/actions
2. By default, "All workflows" are selected (on the left).  Select the
   specific Action for this job (as of this commit, "GitHub Action
   Schedule").
3. Click the "Run workflow" button on the right.
4. Select the branch you want to run it against.
5. Click "Run workflow".

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>