fabtests/configure: fix --with-ze=no option
Add a conditional to only add the --with-ze path to the cflags if it was specified.
Otherwsise, no/ will get added to the flags and make will complain later that the directory does not exist.

Also add a warning/error if ze was explicitly requested but not found.

Also changes define to AC_DEFINE_UNQUOTED to always declare HAVE_LIBZE
and changes ze file check from #ifdef to #if

Cherry-picked from commit 611cc12ef59fcec395c6411dae7249a23471a4af

Signed-off-by: aingerson <alexia.ingerson@intel.com>