prov/cxi: Fix flow control drop count
There is an errata where if a PtlTE is disabled by hardware, the drop
count will miss an increment. This has required software to start the
flow control drop count at -1. But, if software disables the PtlTE, the
errata does not apply. This means software should start the flow control
drop count at 0.

cxip_recv_pte_cb() is updated to detect this case by checking the event
sc_nic_auto field to determine if software or hardware disabled the
PtlTE.

NETCASSINI-2586

Signed-off-by: Ian Ziemba <ian.ziemba@hpe.com>