[v1.12.x] fabtests: Add -K Flag For Benchmark Tests
Currently for fabtests, we do not have fork tests that
call fork to create a child process when running the test.
Given the fact that plenty of applications uses fork
regularly, we need to enable this option.

This commit adds the "-K" flag for the tests under
fabtests/benchmarks directory. With this option, sender
and receiver will both call fork to create a child process after
endpoint was initialized. The child process will do nothing

Signed-off-by: Ao Li <aolia@amazon.com>
Signed-off-by: Wei Zhang <wzam@amazon.com>

(cherry picked from 0a5c76689)