core/mem_monitor: return error when memory minitor failed to start
Currently in ofi_monitors_add_cache(), if a minitor failed
to start and the error code is -FI_ENOSYS, the error is ignored.

This behavior was to handle the situation that device memory
monitors cannot start if the device memory was not
configured. However, it has an unintended consequence:
if a system memory monitor failed to start, the error was ignored.
As a result, the MR cache was initialized successfully, but it
does not support system memory.

To address this issue, this PR made the following change to
ofi_monitors_add_cache():

If a type of memory is not initialized, do not check whether
the monitor can be started.

If a type of memory is initialized, and caller specified
a monitor, and the monitor failed to started, return error.
(the -FI_ENOSYS error will not be ignored)

Signed-off-by: Wei Zhang <wzam@amazon.com>