prov/efa: adjust the timing of clearing deferred memory registration list
Because MR cache, currently memory de-registration is deferred: the
de-registration did not happen when application returned the memory to
the OS, but happens when the deferred memory registration list was cleaned
(by calling ofi_mr_cache_flush).

Because memory de-registration is deferred, the memory is held by efa
kernel module, causing memory leak. This patch call ofi_mr_cache_flush
every time ibv_reg_mr is called to address the issue.

It also remove the old approach which is to call ofi_mr_cache_flush
when cached memory registration is multiple of 512, because the old
appraoch is not frequent enough, and can cause system to ran out of
memory.

Signed-off-by: Wei Zhang <wzam@amazon.com>