prov/tcp: Fix race when EP is disabled during connection handshake
tcpx_ep_disable() must always be executed while holding ep->lock.
There are a few places in the CM code where tcpx_ep_disable() is
called without ep->lock, which could cause memory corruptions if
the same endpoint is accessed in a different context at the same
time (e.g., call to fi_shutdown()).

I sometimes see Libfabric to SIGSEGV in ofi_cq_readfrom() when a
connection establishment is in progress but the remote peer exits.
Sadly, I cannot reproduce the issue reliably but I believe it's
because of the race described in the current patch.

I suspect the bug was introduced in commit
30b93ae025fcfe88f0441f155d487957d1c8978a("prov/tcp: Rework CM state
machine to fix lock inversion").

Signed-off-by: Sylvain Didelot <sdidelot@ddn.com>