prov/tcp: Increase default inject size
The default tcp inject size (64 bytes) is the exact size of the
standard rxm header.  This results in the rxm direct send path
being disabled always.

Increase the default inject size to allow at least a smidgen
of application data to fit in the buffer.

This change has the side effect of allowing rxm to use the
tcp tagged APIs for messages larger than 0 bytes.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>