core: Allow provider to modify the default fi_info attributes
The util_prov references a list of provider fi_info structures.
Those structures are compared against user hints to determine
if a provider can meet the requested user requirements.  For
example, the provider fi_info will have caps set that are
supported, as well as maximum limits, such as the tx_attr::size
value.

However, it's often desirable for the provider to return settings
that are below the max limits.  E.g. the verbs provider does this in
its internal getinfo processing.

Add a callback to struct util_prov that allows a provider to
modify the default fi_info attributes after the base fi_info
attributes are duped.

Update the tcp provider to use the callback to change the
rx/tx sizes to smaller default values, based on whether a
shared receive context is used.

Application specified attributes will still override the defaults.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>