core: Add fastlock_held debug check
Some functions require that a lock be held when they are
called.  Provide a debug check that can verify this.  This
can replace comments in the code that the lock is held with
actual code that documents the requiremen and checks for it.

Update obvious places in the code where locks must be held
to assert on fastlock_held.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>