prov/tcp: Write CONNECTED event to EQ in tcpx_ep_enable()
Fix TODO in the code.  We want to delay writing that an
endpoint has been CONNECTED to the EQ until after the EP
has been fully transitioned.  Otherwise, if an error occurs,
we'll have signaled the the EP is connected, but force it
into an error state immediately afterward.

We also don't want another thread to read the event and
start trying to use it prior to it being fully setup.

Signed-off-by: Sylvain Didelot <sdidelot@ddn.com>