prov/rxm: Combine TX and TX_SAR buffer pools together
RxM maintains separate buffer pools for Tx and SAR buffers.
The result is that both can grow to the size of the Tx
attribute (default 16k), and the regions are registered
with the core provider.  Because only up to tx buffers may
be posted on the endpoint, this results is significant
overhead and memory footprint.

Start reducing the memory footprint by combining these
pools into one.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>