prov/cxi: Introduce RXC reenable state
The current state machine did not properly account for appends in the
queue while the PtlTE gets disabled. Once disabled and before the search
and delete command is issue, LPE can process appends which may match
against unexpected headers which frees LEs. Then, when the search and
delete command is issued, the onloaded ULE count is zero and the
provider would abort stating SW EPs are needed. But, this is incorrect
since a PUT_OVERFLOW event may have occurred against the user receive
buffer which freed an LE.

To fix this, a new state, RXC_ONLOAD_FLOW_CONTROL_REENABLE, is
introduced. This state means that the RXC should always attempted to be
reenabled. Now if a PUT_OVERFLOW for a user buffer occurs in between the
PtlTE disable and when the search and delete event is issued, the RXC
state is updated to RXC_ONLOAD_FLOW_CONTROL_REENABLE and the provider
will not abort if the onloaded ULE count is zero.

NETCASSINI-2653

Signed-off-by: Ian Ziemba <ian.ziemba@hpe.com>