prov/tcp: Report full size of available message in dyn recv callback
The tcp provider reports back the size of the data that has been
read from the socket when invoking the dynamic receive callback.
This works okay when the data includes the rxm header, which
carries the message size at that layer.  But when the tcp
tagged API is used, there isn't an rxm header.  As a result, rxm
uses the size being reported to create a fake received header.

With the current structure, the dynamic receive callback will
always report a size of 0, since only the tcp head will have
been read.  It turns out that rxm only uses the size when
tagged messages are being reported.

Update tcp to return the size of the messages, so that rxm can
record it properly.

It is a mystery how any CI passed with this error, but it did.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>