fabtests: Fix data verification setup
Data verification maintains an iteration count, which is used
as a seed into the data format.  This works fine for most
fabtests, where the client and server are continually restarted.
However, it fails with ubertest, which allows the server to
run continuously for multiple tests.  In the ubertest case,
the second test will fail the data verification check because
the client and server states are out of sync.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>