prov/rxm: Fix initialization of atomic headers
The base rxm header carries the size of the data beyond
the standard rxm header itself (ctrl + op headers).  However
the atomic request and response set the size equal to the
total amount of data being transmitted.  This breaks when
receive side bounce buffers are removed, as the amount of
data expected no longer matches what was sent.

Set the pkt.hdr.size to the correct value, with the base
pkt.hdr size excluded.

We can argue that rxm should report the real size of the
message, rather than just the data len size, but this is
how the current protocol is defined.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>