prov/cxi: Define CXI counter extensions
The flag FI_CXI_CNTR_WB is used to signify to the provider a counter
writeback needs to occur at the completion of the associated deferred
work queue operation.

FI_CXI_COUNTER_OPS can be used with fi_open_ops() on a counter to get a
pointer to struct fi_cxi_cntr_ops. Clients can use the functions defined
in the struct fi_cxi_cntr_ops to update the counter writeback buffer
and/or get the counter MMIO region.

Inline functions are provider to support counter writeback buffer and
MMIO region processing.

Signed-off-by: Ian Ziemba <ian.ziemba@hpe.com>