prov/rxm: Add lock to fast inject path to silence false assert
The inject fast paths are only used where thread safety is not
required.  Those paths can fall through to an emulate inject
path, which needs to allocate a transmit buffer from the tx
pool.  There are assertions in that path that the ep lock is
held.  However, even in the debug build, the fast inject path
skips trying to acquire any lock.

Because if thread safety is diabled, the ep lock/unlock functions
will be no-ops.  We can just acquire/release the ep lock, which
fixes the assertion.  (We want to keep the assertion to detect
real locking issues).

This change will also enable a follow up patch to always use
the fast inject path over the 'slow' inject path.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>