prov/cxi: Flush appends before onloading ULEs
When the PtlTE is disabled, if there are any appends in the command
queue, LPE still processes the appends. This issue with this is that
during onloading, cxil_pte_status() is called to read all the ULE remote
offsets. cxil_pte_state() could be reading the ULE list while hardware
is matching appends against the ULE list. This can lead to incorrect
values being returned in cxil_pte_state().

To resolve this issue, before onloading occurs, the command queue is
flushed of all pending append commands. This, in combination of the
PtlTE being disabled, ensures the ULE list is stable. The flush is
implemented by a search command which will match zero ULEs.

NETCASSINI-2682

Signed-off-by: Ian Ziemba <ian.ziemba@hpe.com>