prov/shm: change shm addressing to only use global ep index
Add global ep_idx to create unique EP names. This eliminates the need for
domain and ep indices and also resolves issues if the same process creates
separate fabrics since they will both be created with a domain and ep index
of 0 causing a failure on the second fi_enable call because the name is
already in use.

The global index is protected by the already existing ep_list_lock since it
serves a similar purpose of tracking the global EPs of the process.

Signed-off-by: aingerson <alexia.ingerson@intel.com>