prov/efa: make rdm_peer part of efa_av_entry
This commit make rdm_peer as a pointer in efa_av_entry.

To facilitate this change. rdm_peer is not allocated
when efa_av_entry is allocated.

For simplicity, a bufpool is used for allocating
rdm_peer, replacing an array, which eliminates the
need for function efa_peer_resize()

Also fix up the rdm_peer release/reset functions
while we're here.

Signed-off-by: Robert Wespetal <wesper@amazon.com>
Signed-off-by: Wei Zhang <wzam@amazon.com>