prov/rxm: Handle error completions properly
Do not generate CQ entries that are returned to the
application for transfers that are associated with
internal sends.  Provide error handling for missing
transfer states.

To prepare this patch a full analysis of the rxm tx/rx
states was done.  Minor code cleanups occur around the
state handling locations.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>