prov/efa: store pointer to efa_conn in efa_av_entry
Add this pointer so we can access the address handle directly and so we
have a pointer to the qkey for this address vector entry.

Once DGRAM is refactored to use util_av this should be changed to store
efa_conn in the efa_av_entry instead of just the pointer.

Signed-off-by: Robert Wespetal <wesper@amazon.com>
Signed-off-by: Wei Zhang <wzam@amazon.com>