prov/rxm: Handle cleanup of peer and fi_addr independently
When establishing a loopback connection, it's possible for the
rxm_cmap_handle fi_addr and peer fields to both be valid.
However, the rxm_conn_handle_notify() path assumes that only
one is valid.  The result is that the cmap->handles_av[]
can be left with a stale entry referencing a freed handle.

Check to see if the handle's fi_addr is valid directly and
remove it from the av prior to freeing the handle.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>