prov/rxm: Fix check for tcp core provider in use_srx
rxm_use_srx() checks if the core provider is tcp and enables
shared rx if so.  The check looks at the provider name.  In
case hints are given, it only checks the leftmost part of the
name string.  Expand the check so that "ofi_rxm;tcp" will
properly report that tcp is the core provider.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>