core/bsock: Define buffered socket abstraction
Introduce a new abstraction called buffered sockets.  The
purpose behind the buffered socket is to isolate the use
of send and receive staging buffers.  That will simplify
changes to the algorithm for when staging buffers are used.

The initial algorithm uses the send side staging buffer
only when the kernel send queue is full.  For small transfers,
the data is copied into the staging buffer, allowing multiple
transfers to coalesce before being sent.  No additional delay
in send operations is introduced.

The receive side staging buffer is used to prefetch data
when the current receive operation is less than half the
staging buffer.  Otherwise, data is received directly into
the buffer from the kernel.

The buffered socket requires that the user provide it a
usable tcp socket.  The abstraction only implements send
and receive operations, and is not a complete abstraction
over the socket.  The user is responsible for the management
of the socket state.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>