prov/tcp: Use rem_len to track receive bytes still needed
TCP uses a generic xfer_entry structure to track send and
receive operations.  For sends, a rem_len field indicates
the number of bytes left to transfer.  For receives, the
field is set in a couple of places, but it's use is
questionable.

Expand the use of rem_len to track receive bytes left to
read from the socket.  This makes the rem_len value
consistent for both send and receive paths.  Update the
dynamic receive buffer path to reflect the change in
semantics.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>