prov/tcp: Make msg response handling align with recv handling
In tcpx_op_msg, if the received message is an ack, we call
tcpx_handle_msg_resp, followed by returning EAGAIN.  In other
cases where handling a receive message, we set a callback to
handle the received data.  The latter allows handling multiple
messages in a loop.  Response message handling breaks that
loop unnecessarily and results in a different execution flow
from what might be expected.

One of the results of the response handling flows is that
we call tcpx_rx_setup with NULL parameters in order to reset
the receive message state.  A follow on patch will adjust
rx_setup, and it's cleaner code to assume that non-null values
are passed into rx_setup.

When dealing with message responses, the incoming receive
should match with the original transmit.  Set a callback to
handle the response in the normal receive handling loop.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>