prov/tcp: Optimize ofi_byteq_read/writev
The offset for writev is always passed in as 0.
Remove the parameter until there's a need
for it.  Without the offset, we can optimize the copy in/out
routines.

Add a check to readv to see if we can use byteq_read
directly and avoid the extra loops handling an iov
with only 1 entry.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>