prov/tcp: Reset for next receive only after completing current
Receive processing is reset in 2 different places.  One is after
reading the header of the current message being processed.  That
resets the read header length, while the header data itself
remains available as the current message.  The second place is
when the rx_entry allocated to receive the message data is freed.

Only reset the endpoint's receive message state when we're done
receiving all data for the message.  This keeps the rx state
consistent.

Perform code cleanup in the touched functions to improve
readability.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>