prov/tcp: Add call to reset the current rx message
The full rx message state is reset in 2 different
locations.  The first part of the state is reset after
reading the header for the received message.  The second
part is after the rx message itself is freed.

This causes issues as the code will show that there's a
currently active message, but that the header has been
reset.  Other optimizations, such as send coalescing, end
up exposing this issue, which shows up as the receiving
processing the same message header multiple times.

Define a single function to reset the full rx message
state.  Future patches will call this to reset the rx
state only after the current rx message has completed
reading all data from the underlying socket.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>