prov/efa: add rdm_peer field to efa_reverse_av
Store a pointer to the rdm_peer for the RDM endpoint in reverse_av,
so that the upper layer can find rdm_peer when polling the completion
queue. The upper layer also needs to do additional validation to
determine queue pair reuse, so it's not appropriate for this hash table
to return an fi_addr_t.

Signed-off-by: Robert Wespetal <wesper@amazon.com>