prov/tcp: Store a reference to the active transmit with the ep
Add a cur_tx_entry pointer to reference the active transmit
entry, similar to the cur_rx_entry.  This avoids needing
to look at the head of the tx_queue, except when checking
for new entries.  Additional patches will expand how
cur_tx_entry is used.

As a result of this change the tx_queue only maintains a list
of pending (non-active) transfers.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>