prov/tcp: Move rem_len from xfer_entry to the endpoint
Every transfer entry maintains a remaining length size
indicating the amount of data left to send or receive.
However, messages are processed serially.  We only need
to maintain one rem_len value for sends and one for
receives.  Both values can be initialized to the message
size when reading or writing the message to the socket.

This reduces the size of the xfer_entry structure
slightly and simplifies the code.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>