prov/tcp: Remove tcpx_rx_setup
With the previous simplifications, tcpx_rx_setup is now a 2-line
function called from 3 places.  Remove it and set the fields
directly.

As part of the change, modify the rx_process_fn to take the
endpoint as input, rather than the rx_entry.  The functions
will always operate on the current rx entry.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>