prov/tcp: Free cur_xx_entry in ep_disable
Free the cur_tx_entry and cur_rx_entry when flushing
the transmit queues in the tcpx_ep_disable path.  This
fixes an assertion that all entries have been returned
to the buffer pools.

If an error for the cur_rx_entry has already been reported,
we need to clear the cur_rx_entry to avoid ep_disable
from reporting a second completion or accessing the freed
rx_entry.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>