prov/tcp: Simplify tcpx_progress_rx
Simplify receive side message progress flow.  The receiver
should be in 1 of 3 basic states.  The first state is
reading the base header.  Once the base header has been
read, we can read the payload offset field to determine
if additional header space is present.

There is a bug in that logic where it checks to see if
the amount of header space read is >= the base header.
Because reading the header is handled in a loop it's
possible for that check to be true multiple times in
the progress rx loop.  For example, after reading the
base header it is true.  That's followed by reading
additional header space in another progress loop,
which can execute the if statement again.

Update the check to only process the base header the
first time it is read.

The second state is to read the remaining headers, if
any.  That is handled by the same receive header call.

The final state is the processing of the message, including
reading any related data.  For that we make use of the
current rx handler() callback.  After receiving the headers,
we simply set the handler callback to one of the start_op
functions.  The start_op calls can still adjust the handler
callback to redirect the progress loop to another receive
callback.  But this simplifies the overall code flow.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>