fabtest: Continue to drive progress when syncing tests
When using OOB sockets to sync the client and server during
tests, we need to continue to drive progress.  Otherwise,
whichever side is waiting to receive data may stall, causing
both sides to spin.

This problem shows up running ubertest over an optimized tcp
provider.  When the tcp optimizations on, send operations may
be coalesced into a provider buffer when the socket is busy.
Once the data has been copied into the buffer, from the
viewpoint of the application, the operation is complete
(FI_INJECT_COMPLETE semantics).  TCP generates a cq event,
which the test reads, followed by blocking on the socket.

The data itself was never queued on the socket.  So without
the test driving progress, both sides end up stalling.

Fix this by changing how the OOB sync works.  Modify the socket
to nonblocking.  Spin until all data is sent/received.  And
drive progress on the Tx CQ in the spin loop.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>