prov/tcp: Fail try_wait if there's prefetch rx data available
If the application uses blocking calls, they will call fi_try_wait
prior to blocking.  However, if data has been queued in the prefetch
buffer the message that the app may be waiting for has already
been read from the socket.  We need to process the data in this
case prior to blocking or the application may hang.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>