prov/efa: introducing extra request "connid header"
This patch introduces a new extra request "need connid" to
efa rdm protocol version 4. This extra request is assigned
ID 3.

An EFA rdm endpoint that want to and can understand
connection ID in packet header, which turn this flag on
in the HANDSHAKE packet's extra_info field.

Connection ID (connid) is a 4-bytes random integer identifies
an endpoint. Including connection ID in a packet's header allows
the receiver of a packet to identify sender of the packet.
The reason it is needed is because device only report GID and QPN
of a received packet, while different send can have same GID + QPN.

EFA uses QKEY as the connection ID. This commit consists of
following changes:

1. An extra feature flag:

    RXR_EXTRA_REQUEST_CONNID_HEADER

2. Let an rdm end point claim support of it.

3. A function rxr_peer_need_connid() to determine
   whether a peer has this request.

4. A universal flag RXR_PKT_CONNID_HDR all packet types will use
   to indicate the presence of connid in packet header.

Signed-off-by: Wei Zhang <wzam@amazon.com>